requirements.txt
google-generativeai
openai
PyPDF2